app/requirement.txt
beautifulsoup4==4.6.0
Flask==0.12.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
gevent==1.2.1
Pillow==4.1.1
pyquery==1.2.13
PySocks==1.5.6
requests==2.10.0
shadowsocks==2.8.2
SQLAlchemy==1.2.3
zbarlight==1.2
gunicorn
psutil==5.4.3
